Improve scalability in testimonials
The testimonial has been re-styled in the form of an automatic Slideshow.

Regarding the database, four new fields has been added in the testimonials table:
- quotes: all the translations of the quote
- occupation (Just to be displayed in the view)
- company (Just to be displayed in the view)
- user_id: references to the users table

A new observer class has been created when retrieving testimonials to serialize the locales of every quote from each testimonial registered,
since these are populated as a JSON string.

Some needed methods have been created in the Testimonial model to retrieve the data ready to be displayed inside the view.

The fillables of the Testimonial model has been updated according to the changes.

A new method to retrieved the last user inserted has been created in order to be able to bind a testimonial record to an existing user
while seeding the database.

Defined testimonial seedings and migrations.

A new field has been added in the users table:
- Avatar: comprises a profile picture of the user and it is used in the testimonials.

A new slider class extending from the slider factory has been created as to be applied in the testimonial section.

Remove unused files.